+++ Source/Native/Engine/Public/Engine/Rendering/RenderGraph.h
@@ -29,12 +29,12 @@ namespace march
         void SetBuffer(const std::string& name, GfxBuffer* value) { m_Context->SetBuffer(name, value); }
         void SetBuffer(int32_t id, GfxBuffer* value) { m_Context->SetBuffer(id, value); }
 
-        void DrawMesh(GfxMeshGeometry geometry, Material* material, int32_t shaderPassIndex) { m_Context->DrawMesh(geometry, material, shaderPassIndex); }
-        void DrawMesh(GfxMeshGeometry geometry, Material* material, int32_t shaderPassIndex, const DirectX::XMFLOAT4X4& matrix) { m_Context->DrawMesh(geometry, material, shaderPassIndex, matrix); }
-        void DrawMesh(GfxMesh* mesh, uint32_t subMeshIndex, Material* material, int32_t shaderPassIndex) { m_Context->DrawMesh(mesh, subMeshIndex, material, shaderPassIndex); }
-        void DrawMesh(GfxMesh* mesh, uint32_t subMeshIndex, Material* material, int32_t shaderPassIndex, const DirectX::XMFLOAT4X4& matrix) { m_Context->DrawMesh(mesh, subMeshIndex, material, shaderPassIndex, matrix); }
-        void DrawMesh(const GfxSubMeshDesc& subMesh, Material* material, int32_t shaderPassIndex) { m_Context->DrawMesh(subMesh, material, shaderPassIndex); }
-        void DrawMesh(const GfxSubMeshDesc& subMesh, Material* material, int32_t shaderPassIndex, const DirectX::XMFLOAT4X4& matrix) { m_Context->DrawMesh(subMesh, material, shaderPassIndex, matrix); }
+        void DrawMesh(GfxMeshGeometry geometry, Material* material, size_t shaderPassIndex) { m_Context->DrawMesh(geometry, material, shaderPassIndex); }
+        void DrawMesh(GfxMeshGeometry geometry, Material* material, size_t shaderPassIndex, const DirectX::XMFLOAT4X4& matrix) { m_Context->DrawMesh(geometry, material, shaderPassIndex, matrix); }
+        void DrawMesh(GfxMesh* mesh, uint32_t subMeshIndex, Material* material, size_t shaderPassIndex) { m_Context->DrawMesh(mesh, subMeshIndex, material, shaderPassIndex); }
+        void DrawMesh(GfxMesh* mesh, uint32_t subMeshIndex, Material* material, size_t shaderPassIndex, const DirectX::XMFLOAT4X4& matrix) { m_Context->DrawMesh(mesh, subMeshIndex, material, shaderPassIndex, matrix); }
+        void DrawMesh(const GfxSubMeshDesc& subMesh, Material* material, size_t shaderPassIndex) { m_Context->DrawMesh(subMesh, material, shaderPassIndex); }
+        void DrawMesh(const GfxSubMeshDesc& subMesh, Material* material, size_t shaderPassIndex, const DirectX::XMFLOAT4X4& matrix) { m_Context->DrawMesh(subMesh, material, shaderPassIndex, matrix); }
         void DrawMeshRenderers(size_t numRenderers, MeshRenderer* const* renderers, const std::string& lightMode) { m_Context->DrawMeshRenderers(numRenderers, renderers, lightMode); }
 
         void ResolveTexture(GfxTexture* source, GfxTexture* destination) { m_Context->ResolveTexture(source, destination); }
